move number input helpers to number handler
